requirements.txt
fastapi
uvicorn
motor
pymongo
openai
textblob
pydantic
python-dotenv
streamlit
requests
